+++ NetIdMobileSdk/src/androidTest/java/de/netid/mobile/sdk/NetIdPermissionTest.kt
@@ -27,7 +27,6 @@ class NetIdPermissionTest: NetIdServiceListener {
 
     companion object {
         private const val clientId = "082531ba-1b22-4381-81b1-64add4b85b8a"
-        private const val host = "broker.netid.de"
         private const val redirectUri = "https://netid-sdk-web.letsdev.de/redirect"
         private const val claims = "{\"userinfo\":{\"email\": {\"essential\": true}, \"email_verified\": {\"essential\": true}}}"
 
@@ -37,7 +36,7 @@ class NetIdPermissionTest: NetIdServiceListener {
     @Before
     public fun setup() {
         assertEquals("de.netid.mobile.sdk.test", appContext.packageName)
-        val netIdConfig = NetIdConfig(host, clientId, redirectUri, claims)
+        val netIdConfig = NetIdConfig(clientId, redirectUri, claims)
         NetIdService.addListener(this)
 
         NetIdService.initialize(netIdConfig, appContext)