+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/ui/AuthorizationLoginFragment.kt
@@ -27,28 +27,22 @@ import android.view.ViewGroup
 import android.widget.LinearLayout
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.content.res.ResourcesCompat
-import androidx.core.content.res.ResourcesCompat.getDrawable
 import androidx.fragment.app.Fragment
 import com.google.android.material.button.MaterialButton
 import de.netid.mobile.sdk.R
 import de.netid.mobile.sdk.api.NetIdAuthFlow
-import de.netid.mobile.sdk.databinding.FragmentAuthorizationHardBinding
+import de.netid.mobile.sdk.databinding.FragmentAuthorizationLoginBinding
 import de.netid.mobile.sdk.model.AppIdentifier
 
-class AuthorizationHardFragment(
+class AuthorizationLoginFragment(
     private val listener: AuthorizationFragmentListener,
     private val appIdentifiers: List<AppIdentifier>,
     private val authorizationIntent: Intent,
-    private val authFlow: NetIdAuthFlow,
     private val headlineText: String = "",
     private val loginText: String = "",
     private val continueText: String = ""
 ) : Fragment() {
-    companion object {
-        private const val netIdScheme = "scheme"
-    }
-
-    private var _binding: FragmentAuthorizationHardBinding? = null
+    private var _binding: FragmentAuthorizationLoginBinding? = null
 
     // This property is only valid between onCreateView and onDestroyView.
     private val binding get() = _binding!!
@@ -68,7 +62,7 @@ class AuthorizationHardFragment(
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentAuthorizationHardBinding.inflate(inflater, container, false)
+        _binding = FragmentAuthorizationLoginBinding.inflate(inflater, container, false)
         return binding.root
     }
 
@@ -83,23 +77,18 @@ class AuthorizationHardFragment(
     }
 
     private fun setupStandardButtons() {
-        var currentAppName = "App"
-        if (context?.applicationInfo?.name.toString().uppercase() != "NULL") {
-            currentAppName = context?.applicationInfo?.name.toString()
-        }
-//        val continueString = getString(R.string.authorization_hard_continue, currentAppName)
-//        binding.fragmentAuthorizationButtonClose.text = continueString.uppercase()
         binding.fragmentAuthorizationButtonClose.setOnClickListener {
             listener.onCloseClicked()
         }
         binding.fragmentAuthorizationButtonAgreeAndContinue.setOnClickListener {
+            authorizationIntent.setPackage("com.android.chrome")
             resultLauncher.launch(authorizationIntent)
         }
     }
 
     private fun setupAppButtons() {
         // If there are no apps installed, display a standard button to enable app2web flow
-        if (appIdentifiers.size == 0) {
+        if (appIdentifiers.isEmpty()) {
             binding.fragmentAuthorizationButtonAgreeAndContinue.visibility = View.VISIBLE
         }
 
@@ -115,7 +104,11 @@ class AuthorizationHardFragment(
 
     public fun createButton(appIdentifier: AppIdentifier): MaterialButton {
         val appButton = MaterialButton(requireContext(), null, com.google.android.material.R.attr.borderlessButtonStyle)
-        val continueString = getString(R.string.authorization_hard_continue, appIdentifier.name)
+        val continueString = if (loginText.isEmpty()) {
+            getString(R.string.authorization_login_continue, appIdentifier.name)
+        } else {
+            String.format(loginText, appIdentifier.name)
+        }
         val resourceId =
             context?.resources?.getIdentifier(appIdentifiers[0].typeFaceIcon, "drawable", requireContext().opPackageName)
         appButton.icon = resourceId?.let {
@@ -142,9 +135,7 @@ class AuthorizationHardFragment(
         appButton.layoutParams = layoutParams
 
         appButton.setOnClickListener {
-            resultLauncher.launch(authorizationIntent)
-            listener.onAppButtonClicked(appIdentifier)
-            openApp(appIdentifier.android.verifiedAppLink)
+            openApp(appIdentifier)
         }
 
         return appButton
@@ -155,15 +146,15 @@ class AuthorizationHardFragment(
         super.onDestroyView()
     }
 
-    private fun openApp(verifiedAppLink: String) {
-        val authIntent = authorizationIntent.extras?.get("authIntent")as Intent
-        val authUri = authIntent.data as Uri
-        val uri = authUri.toString().replaceBefore("?", verifiedAppLink)
-
-        val intent = Intent(Intent.ACTION_VIEW, Uri.parse(uri))
-//        intent?.putExtra(netIdScheme, context?.applicationInfo?.packageName)
-        intent.let {
-            context?.startActivity(intent)
+    private fun openApp(app: AppIdentifier) {
+        authorizationIntent.extras?.apply {
+            val authIntent = getParcelable<Intent>("authIntent") ?: return@apply
+            val authUri = authIntent.data as Uri
+            val uri = authUri.toString().replaceBefore("?", app.android.verifiedAppLink)
+            authIntent.setPackage(app.android.applicationId)
+            authIntent.data = Uri.parse(uri)
+            putParcelable("authIntent", authIntent)
         }
+        resultLauncher.launch(authorizationIntent)
     }
 }