+++ app/src/main/java/de/netid/mobile/sdk/example/MainActivity.kt
@@ -37,9 +37,10 @@ class MainActivity : AppCompatActivity(), NetIdServiceListener {
 
     companion object {
         private const val clientId = "082531ba-1b22-4381-81b1-64add4b85b8a"
-        private const val host = "broker.netid.de"
         private const val redirectUri = "https://netid-sdk-web.letsdev.de/redirect"
         private const val claims = "{\"userinfo\":{\"email\": {\"essential\": true}, \"email_verified\": {\"essential\": true}}}"
+        private val permissionLayerConfig = null
+        private val loginLayerConfig = null
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -84,7 +85,7 @@ class MainActivity : AppCompatActivity(), NetIdServiceListener {
     }
 
     private fun setupNetIdConfig() {
-        netIdConfig = NetIdConfig(host, clientId, redirectUri, claims)
+        netIdConfig = NetIdConfig(clientId, redirectUri, claims, permissionLayerConfig, loginLayerConfig)
         NetIdService.addListener(this)
     }
 
@@ -223,6 +224,9 @@ class MainActivity : AppCompatActivity(), NetIdServiceListener {
     override fun onAuthenticationFinished(accessToken: String) {
         appendLog("netID service authorized successfully\nAccess Token:\n$accessToken")
         serviceState = ServiceState.AuthorizationSuccessful
+        if ((this::bottomDialogFragment.isInitialized) && (bottomDialogFragment.isAdded)) {
+            bottomDialogFragment.dismiss()
+        }
         updateElementsForServiceState()
     }